Suppress non-ASCII characters in review request fields
Non-ASCII characters in the summary or description fields would cause
ReviewBoard to reject the request, so we now replace them with '?' characters.
Also fixed the exception code so that it doesn't try to put the entire error
response in the exception string if it isn't JSON. This caused failures in
the Mercurial error output functions when the response was a large HTML
document.